IO: move helper types into the io_types module
Create sherpa.astro.io.io_types as a place to hold the new types
which are planned to improve the typing of the interface between
the I/O layer and the I/O backends.
