+++ docs/overview/astro_io_io_types.rst
@@ -1,0 +1,22 @@
+***********************************
+The sherpa.astro.io.io_types module
+***********************************
+
+.. currentmodule:: sherpa.astro.io.io_types
+
+.. automodule:: sherpa.astro.io.io_types
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree: api
+
+      Column
+      HeaderItem
+      TableHDU
+
+Class Inheritance Diagram
+=========================
+
+.. inheritance-diagram:: Column HeaderItem TableHDU
+   :parts: 1
