+++ docs/overview/utilities.rst
@@ -56,6 +56,7 @@ Reference/API
    io
    astro
    astro_io
+   astro_io_io_types
    astro_io_wcs
    astro_io_xstable
    astro_utils_xspec
